requirements.txt
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
appscript==1.4.0
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.2
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
contourpy==1.2.1
cryptography==46.0.3
curl_cffi==0.13.0
cycler==0.12.1
distro==1.9.0
docx2pdf==0.1.8
docxcompose==1.4.0
fastapi==0.124.4
fastjsonschema==2.21.2
filelock==3.20.0
Flask==3.1.0
fonttools==4.53.1
frozendict==2.4.7
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.43.0
google-genai==1.52.0
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.12.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_core==5.9.1
kiwisolver==1.4.5
langchain==1.1.0
langchain-core==1.1.0
langgraph==1.0.4
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.10
langsmith==0.4.49
lxml==6.0.2
MarkupSafe==3.0.2
matplotlib==3.9.1
mpmath==1.3.0
multitasking==0.0.12
mysql-connector==2.2.9
mysql-connector-python==9.1.0
narwhals==2.12.0
nbformat==5.10.4
networkx==3.6
numpy==2.0.1
openai==2.11.0
orjson==3.11.4
ormsgpack==1.12.0
packaging==24.1
pandas==2.3.3
pdf2image==1.17.0
pdfminer.six==20251107
pdfplumber==0.11.8
peewee==3.18.3
pillow==10.4.0
platformdirs==4.5.0
plotly==6.5.0
protobuf==6.33.1
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
PyMuPDF==1.26.7
pyparsing==3.1.2
pypdfium2==5.2.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.29.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.1.2
setuptools==80.9.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
starlette==0.50.0
streamlit==1.51.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.1
toml==0.10.2
torch==2.9.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
xxhash==3.6.0
yfinance==0.2.66
zstandard==0.25.0
